Setting up repo for SPICE ingest
