+++ .devcontainer/Dockerfile
@@ -11,8 +11,8 @@ RUN unzip awscliv2.zip
 RUN sudo ./aws/install
 RUN sudo apt-get install -y python3 python3-pip
 
-RUN pip install poetry
-ENV AWS_CONFIG_FILE=/workspaces/SDS-in-a-box/.aws/config
-ENV AWS_SHARED_CREDENTIALS_FILE=/workspaces/SDS-in-a-box/.aws/credentials
+RUN pip install "poetry<2"
+ENV AWS_CONFIG_FILE=/workspaces/sds-data-manager/.aws/config
+ENV AWS_SHARED_CREDENTIALS_FILE=/workspaces/sds-data-manager/.aws/credentials
 
-WORKDIR /workspaces/SDS-in-a-box/
+WORKDIR /workspaces/sds-data-manager/
