+++ sds_data_manager/constructs/efs_construct.py
@@ -1,17 +1,8 @@
 """Configure the EFS."""
 
 import aws_cdk as cdk
-from aws_cdk import (
-    Duration,
-    Environment,
-    aws_iam,
-    aws_lambda,
-)
 from aws_cdk import aws_ec2 as ec2
 from aws_cdk import aws_efs as efs
-from aws_cdk import aws_events as events
-from aws_cdk import aws_events_targets as targets
-from aws_cdk import aws_s3 as s3
 from constructs import Construct
 
 
@@ -109,117 +100,3 @@ class EFSConstruct(Construct):
             path=self.efs_spice_path,
             posix_user=efs.PosixUser(gid="1000", uid="1000"),
         )
-
-
-class EFSWriteLambda(Construct):
-    """Create some Lambdas that write to the EFS file system."""
-
-    def __init__(
-        self,
-        scope: Construct,
-        construct_id: str,
-        code: aws_lambda.Code,
-        env: Environment,
-        vpc: ec2.Vpc,
-        data_bucket: s3.Bucket,
-        efs_construct: EFSConstruct,
-        **kwargs,
-    ) -> None:
-        """Construct the EFS lambdas.
-
-        Parameters
-        ----------
-        scope : Construct
-            Parent construct.
-        construct_id : str
-            A unique string identifier for this construct.
-        code : aws_lambda.Code
-            Lambda code bundle
-        env : Environment
-            Account and region
-        vpc : ec2.Vpc
-            VPC into which to put the resources that require networking.
-        data_bucket : obj
-            The data bucket
-        efs_construct : obj
-            The EFS filesystem construct
-        kwargs : dict
-            Keyword arguments
-
-        """
-        super().__init__(scope, construct_id, **kwargs)
-
-        # Create a role for the EFS Lambda
-        # Grant the Lambda identity role access to the VPC/EFS
-        iam_role_name = "efs-lambda-role"
-        efs_lambda_role = aws_iam.Role(
-            self,
-            iam_role_name,
-            role_name=iam_role_name,
-            managed_policies=[
-                aws_iam.ManagedPolicy.from_aws_managed_policy_name(
-                    "service-role/AWSLambdaBasicExecutionRole"
-                ),
-                aws_iam.ManagedPolicy.from_aws_managed_policy_name(
-                    "service-role/AWSLambdaVPCAccessExecutionRole"
-                ),
-                aws_iam.ManagedPolicy.from_aws_managed_policy_name(
-                    "AmazonS3FullAccess"
-                ),
-                aws_iam.ManagedPolicy.from_aws_managed_policy_name(
-                    "AmazonElasticFileSystemFullAccess"
-                ),
-            ],
-            assumed_by=aws_iam.ServicePrincipal("lambda.amazonaws.com"),
-        )
-
-        # This access point is used by other resources to read from EFS
-        spice_mount_path = "/mnt/spice"
-
-        self.efs_spice_ingest_lambda = aws_lambda.Function(
-            self,
-            "EFSWriteLambda",
-            function_name="efs-write-lambda",
-            # Allow access to the EFS over NFS port
-            allow_all_outbound=True,
-            runtime=aws_lambda.Runtime.PYTHON_3_12,
-            code=code,
-            handler="efs_lambda.lambda_function.lambda_handler",
-            role=efs_lambda_role,
-            description="Lambda that write data to the EFS",
-            # Access to the EFS requires to be within the VPC
-            vpc=vpc,
-            # Mount EFS access point to /mnt/data within the lambda
-            filesystem=aws_lambda.FileSystem.from_efs_access_point(
-                efs_construct.spice_access_point, spice_mount_path
-            ),
-            timeout=Duration.minutes(1),
-            architecture=aws_lambda.Architecture.ARM_64,
-            environment={
-                "EFS_SPICE_MOUNT_PATH": spice_mount_path,
-            },
-        )
-
-        # Trigger lambda on all s3 object creations through
-        # eventbridge
-
-        # Define an EventBridge rule
-        event_rule = events.Rule(
-            self,
-            "EfsWriteLambdaS3EventRule",
-            event_pattern=events.EventPattern(
-                source=["aws.s3"],
-                detail_type=["Object Created"],
-                detail={
-                    "bucket": {"name": [data_bucket.bucket_name]},
-                    "object": {
-                        "key": [
-                            {"prefix": "spice/"},
-                        ]
-                    },
-                },
-            ),
-        )
-
-        # Add the Lambda function as the target for the rule
-        event_rule.add_target(targets.LambdaFunction(self.efs_spice_ingest_lambda))
